requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
dateutils==0.6.12
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
